// src/modules/skills/skills.interface.d.ts
import type { HydratedDocument, Model } from "mongoose";
import type { IImage } from "../../interfaces/index.js";
export interface ISkill {
    name: string;
    slug: string;
    category: string;
    proficiency: number;
    image?: IImage;
    description?: string;
    sortOrder: number;
    isActive: boolean;
    createdAt?: Date;
    updatedAt?: Date;
}
export type TSkillDocument = HydratedDocument<ISkill>;
export interface ISkillModel extends Model<ISkill> {
}
//# sourceMappingURL=skills.interface.d.ts.map